tests(infra): add unit tests to observability interceptor
